Revert "[#7246] ArchivableAdo"
This reverts commit 787d2f7d6b89c67302b6816940879afee2644d0e.
