+++ sormas-backend/src/main/java/de/symeda/sormas/backend/common/AbstractDeletableAdoService.java
@@ -7,7 +7,7 @@ import javax.persistence.criteria.From;
 import javax.persistence.criteria.JoinType;
 import javax.persistence.criteria.Predicate;
 
-public abstract class AbstractDeletableAdoService<ADO extends CoreAdo> extends AdoServiceWithUserFilter<ADO> {
+public abstract class AbstractDeletableAdoService<ADO extends DeletableAdo> extends AdoServiceWithUserFilter<ADO> {
 
 	protected AbstractDeletableAdoService(Class<ADO> elementClass) {
 		super(elementClass);
