+++ sormas-backend/src/main/java/de/symeda/sormas/backend/common/CoreAdo.java
@@ -27,21 +27,19 @@ import de.symeda.auditlog.api.Audited;
  */
 @MappedSuperclass
 @Audited
-public class CoreAdo extends ArchivableAdo {
+public class CoreAdo extends DeletableAdo {
 
-	private static final long serialVersionUID = 6512756286608581221L;
+    public static final String ARCHIVED = "archived";
 
-	public static final String DELETED = "deleted";
+    private boolean archived;
 
-	private boolean deleted;
+    @Column
+    public boolean isArchived() {
+        return archived;
+    }
 
-	public void setDeleted(boolean deleted) {
-		this.deleted = deleted;
-	}
-
-	@Column
-	public boolean isDeleted() {
-		return deleted;
-	}
+    public void setArchived(boolean archived) {
+        this.archived = archived;
+    }
 
 }
