+++ sormas-backend/src/main/java/de/symeda/sormas/backend/labmessage/LabMessage.java
@@ -24,13 +24,13 @@ import de.symeda.sormas.api.person.Sex;
 import de.symeda.sormas.api.sample.PathogenTestResultType;
 import de.symeda.sormas.api.sample.SampleMaterial;
 import de.symeda.sormas.api.sample.SpecimenCondition;
-import de.symeda.sormas.backend.common.CoreAdo;
+import de.symeda.sormas.backend.common.DeletableAdo;
 import de.symeda.sormas.backend.sample.Sample;
 import de.symeda.sormas.backend.user.User;
 
 @Entity(name = "labmessage")
 @Audited
-public class LabMessage extends CoreAdo {
+public class LabMessage extends DeletableAdo {
 
 	public static final String TABLE_NAME = "labmessage";
