+++ sormas-backend/src/main/java/de/symeda/sormas/backend/sample/AdditionalTestService.java
@@ -24,7 +24,7 @@ import de.symeda.sormas.api.sample.AdditionalTestCriteria;
 import de.symeda.sormas.api.utils.SortProperty;
 import de.symeda.sormas.backend.caze.Case;
 import de.symeda.sormas.backend.common.AdoServiceWithUserFilter;
-import de.symeda.sormas.backend.common.ArchivableAdo;
+import de.symeda.sormas.backend.common.DeletableAdo;
 import de.symeda.sormas.backend.common.CriteriaBuilderHelper;
 import de.symeda.sormas.backend.contact.Contact;
 import de.symeda.sormas.backend.event.EventParticipant;
@@ -174,7 +174,7 @@ public class AdditionalTestService extends AdoServiceWithUserFilter<AdditionalTe
 	}
 
 	/**
-	 * Creates a filter that excludes all samples that are {@link ArchivableAdo#deleted} or associated with
+	 * Creates a filter that excludes all samples that are {@link DeletableAdo#deleted} or associated with
 	 * cases that are {@link Case#archived}, contacts that are {@link Contact#deleted}. or event participants that are
 	 * {@link EventParticipant#deleted}
 	 */
