+++ sormas-backend/src/main/java/de/symeda/sormas/backend/sample/PathogenTest.java
@@ -40,14 +40,14 @@ import de.symeda.sormas.api.sample.PCRTestSpecification;
 import de.symeda.sormas.api.sample.PathogenTestReferenceDto;
 import de.symeda.sormas.api.sample.PathogenTestResultType;
 import de.symeda.sormas.api.sample.PathogenTestType;
-import de.symeda.sormas.backend.common.CoreAdo;
+import de.symeda.sormas.backend.common.DeletableAdo;
 import de.symeda.sormas.backend.disease.DiseaseVariantConverter;
 import de.symeda.sormas.backend.infrastructure.facility.Facility;
 import de.symeda.sormas.backend.user.User;
 
 @Entity
 @Audited
-public class PathogenTest extends CoreAdo {
+public class PathogenTest extends DeletableAdo {
 
 	private static final long serialVersionUID = 2290351143518627813L;
