+++ sormas-backend/src/main/java/de/symeda/sormas/backend/sample/PathogenTestService.java
@@ -47,7 +47,7 @@ import de.symeda.sormas.api.utils.SortProperty;
 import de.symeda.sormas.backend.caze.Case;
 import de.symeda.sormas.backend.common.AbstractDeletableAdoService;
 import de.symeda.sormas.backend.common.AbstractDomainObject;
-import de.symeda.sormas.backend.common.ArchivableAdo;
+import de.symeda.sormas.backend.common.DeletableAdo;
 import de.symeda.sormas.backend.common.CriteriaBuilderHelper;
 import de.symeda.sormas.backend.contact.Contact;
 import de.symeda.sormas.backend.event.EventParticipant;
@@ -324,7 +324,7 @@ public class PathogenTestService extends AbstractDeletableAdoService<PathogenTes
 	}
 
 	/**
-	 * Creates a filter that excludes all pathogen tests that are {@link ArchivableAdo#deleted} or associated with
+	 * Creates a filter that excludes all pathogen tests that are {@link DeletableAdo#deleted} or associated with
 	 * cases that are {@link Case#archived}, contacts that are {@link Contact#deleted}. or event participants that are
 	 * {@link EventParticipant#deleted}
 	 */
@@ -336,7 +336,7 @@ public class PathogenTestService extends AbstractDeletableAdoService<PathogenTes
 
 	/**
 	 * Creates a default filter that should be used as the basis of queries in this service..
-	 * This essentially removes {@link ArchivableAdo#deleted} pathogen tests from the queries.
+	 * This essentially removes {@link DeletableAdo#deleted} pathogen tests from the queries.
 	 */
 	public Predicate createDefaultFilter(CriteriaBuilder cb, Root<PathogenTest> root) {
 		return cb.isFalse(root.get(PathogenTest.DELETED));
