+++ sormas-ui/src/test/resources/META-INF/persistence.xml
@@ -10,7 +10,7 @@
 		<class>de.symeda.sormas.backend.action.Action</class>
 		<class>de.symeda.sormas.backend.caze.Case</class>
 		<class>de.symeda.sormas.backend.common.AbstractDomainObject</class>
-		<class>de.symeda.sormas.backend.common.ArchivableAdo</class>
+		<class>de.symeda.sormas.backend.common.DeletableAdo</class>
 		<class>de.symeda.sormas.backend.common.CoreAdo</class>
 		<class>de.symeda.sormas.backend.contact.Contact</class>
 		<class>de.symeda.sormas.backend.epidata.EpiData</class>
